jjb/edgex-go/edgex-go-snap.yaml: downgrade california to beta for now
Until we do more testing we want the california snap builds to be released to beta with the current `grade: devel`.
When we are ready and sure that the snap is more stable, we will re-up the grade to stable, and change the risk here back to candidate. Then finally at some point we will just upgrade one of the candidate releases to stable manually.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>