jjb/many: enable docker builds and remove ubuntu node requirements
This PR enables using the build script for the snap that uses docker to build the snap. Because we're now using docker, we don't need to use specific OS requirements, and we can simplify the job definitions.
Additionally, renaming the PR comment to just '{project-name}-{stream}-verify', as the project-name already specifies that it's for the snap, and on GitHub for example the comment shows up as "edgex-go-snap-snap-master-verfiy", etc.
Lastly, because the build script + docker images now contain the logic necessary to publish the snap using the JENKINS_URL and JOB_NAME environment variables, we can delete the shell/edgex-publish-snap.sh script here as well.
This needs to be merged after https://github.com/edgexfoundry/edgex-go/pull/563

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>