Add new snap build script for snap jobs
This script is meant to build snaps and is combined from the current
files in edgex-go, but generalized to work with any repo with a
snapcraft.yaml in it.

We use this script as the build_script for the default jenkins job
boiler plate for snap job templates so that simply defining the project
and streams for a new repo with appropriately named jobs will enable
snap builds for new repos.

Note that we have to currently store both the Dockerfile and the
entrypoint.sh script inside the edgexfoundry-snapcraft.sh script using
heredocs because we need those things to be available to the job when it
runs and it's easier to store them inside this single script which gets
injected into the job definition rather than try to pull those files
from somewhere else when the job runs.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>