jjb/edgex-go/edgex-go-snap.yaml: drop delhi snap jobs
Delhi is no longer supported as a release, so we can stop building these jobs.

Additionally, the jobs have been failing for some time due in part to a bug with
the packaging that specified the master branch for docker-edgex-mongo, when it
should have specified the delhi branch instead. Rather than fix that bug, we'll
just stop builds for this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>