edgex-templates-integration.yaml: set no security env var
This is for the current set of blackbox tests that don't use security to
hopefully work again, because there was an upstream change in edgex-go which
changed the default behavior of the services to assume they need to run with
security. However, these tests specifically are meant to run without security,
so we need to explicitly disable the security features of EdgeX services.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>