Remove EdgeX custom Ubuntu 18.04 arm64 file
The EdgeX var file for Ubuntu 18.04 arm64 is identical to the
common-packer version, other than the base_image. This routes around
updates and fixes made in common-packer, making upkeep unnecessarily
difficult. The best choice now is to remove the custom file, and only
reintroduce it if EdgeX's requirements deviate from common-packer in
the future.

Issue: LF-Jira RELENG-2751
Signed-off-by: Eric Ball <eball@linuxfoundation.org>