edgexfoundry-snapcraft.sh: use && instead of ; between cmds
The previous behavior would not actually fail the build setup if for example the
SHA's didn't match, but now if the SHA's don't match, sha512sum will exit
non-zero and using && results in the whole command failing.

We also should delete the sha512 file as well, it is not needed during the
build.

Fixes: https://github.com/edgexfoundry/ci-management/issues/517

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>